requirements.txt
asgiref==3.4.1
Django==3.2.9
django-admin-rangefilter==0.8.2
djangorestframework==3.12.4
djongo==1.3.6
numpy==1.21.4
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4

# Bare minimum
django-cms>=3.0

# These dependencies are brought in by django CMS, but if you want to
# lock-in their version, specify them
Django>=1.7

South==1.0.2 # Only needed for Django < 1.7
django-mptt
django-sekizai==0.7
django-classy-tags==0.5
djangocms-admin-style==0.2.2
html5lib==0.999
six==1.3.0

# Optional, recommended packages
Pillow>=2
django-filer
cmsplugin-filer==0.10.1
django-reversion==1.8
